feat(members): show the spend charts on the member profile too
The issue parked Profile -> Usage as a separate dataset, but the numbers are
already readable by any member of the company — /profile/usage?user_id= is not
admin-gated — so keeping them one URL away only made people leave the page they
had opened to ask the question. The heatmap, the five headline numbers, the
per-project split, the agent mix and the daily cost/token curves now render on
/user/:id, over a window the viewer can change.

The panels move to shared/resources/usage/UsageAnalytics, where both pages
render them: each reads its own deferred prop off usePage(), so a page opts in
by serving props under those names and nothing renders until they arrive. The
controller serves them from the same four services, in the same "usage"
deferral group, scoped to the subject and the current company. Unlike
ProfileController#usage it validates `period` rather than forwarding whatever
arrived, so an unknown window falls back to 30 days instead of reaching a
service that cannot answer it.
