+++ app/frontend/pages/Profile/Usage.tsx
@@ -1,87 +1,16 @@
 import { Deferred, Head, router, usePage } from '@inertiajs/react';
-import {
-  Alert,
-  Badge,
-  Box,
-  Divider,
-  Grid,
-  Group,
-  Paper,
-  Select,
-  SimpleGrid,
-  Skeleton,
-  Table,
-  Text,
-  Title,
-  Tooltip,
-} from '@mantine/core';
-import { IconChartBar, IconClock, IconCoin, IconInfoCircle, IconPlayerPlay, IconRoute } from '@tabler/icons-react';
+import { Alert, Badge, Box, Divider, Group, Paper, Select, Skeleton, Table, Text, Title, Tooltip } from '@mantine/core';
+import { IconInfoCircle } from '@tabler/icons-react';
 import { formatDistanceToNow } from 'date-fns';
-import { useMemo } from 'react';
-import {
-  Area,
-  AreaChart,
-  CartesianGrid,
-  Cell,
-  Pie,
-  PieChart,
-  ResponsiveContainer,
-  Tooltip as RechartsTooltip,
-  XAxis,
-  YAxis,
-} from 'recharts';
 
 import { AuthLayout } from 'layouts/AuthLayout';
 
-import { formatCostCents, formatTokens } from 'shared/lib/formatUsage';
-import { CHART_SERIES } from 'shared/theme/chartPalette';
+import { formatTokens } from 'shared/lib/formatUsage';
+import { PERIOD_OPTIONS, UsageAnalytics, type Period } from 'shared/resources/usage/UsageAnalytics';
 import { type SharedProps } from 'shared/ui';
-import { ContributionHeatmap } from 'shared/ui/ContributionHeatmap';
 import { StatusBadge } from 'shared/ui/StatusBadge';
 
 import { ProfileTabs } from './ProfileTabs';
-
-type Period = '7d' | '30d' | '90d' | '1y';
-
-interface ProjectBreakdown {
-  projectId: number | null;
-  projectName: string;
-  sessions: number;
-  costCents: number;
-  tokens: number;
-}
-
-interface SummaryData {
-  totalSessions: number;
-  totalCostCents: number;
-  totalTokens: number;
-  avgCostCentsPerSession: number;
-  workflowsRun: number;
-  projectBreakdowns: ProjectBreakdown[];
-}
-
-interface AgentSessionCount {
-  agentType: string;
-  sessions: number;
-  costCents: number;
-  tokens: number;
-}
-interface AgentActivityData {
-  sessionsByAgent: AgentSessionCount[];
-}
-
-interface CostTokenPoint {
-  date: string;
-  costCents: number;
-  totalTokens: number;
-}
-interface CostTokenData {
-  timeSeries: CostTokenPoint[];
-}
-
-interface HeatmapData {
-  days: { date: string; count: number }[];
-}
 
 interface Session {
   id: number;
@@ -102,23 +31,8 @@ interface Props {
   projectId?: string | null;
   viewerIsSelf: boolean;
   targetUser: { id: number; name: string | null; email: string };
-  summary?: SummaryData;
-  agentActivity?: AgentActivityData;
-  costToken?: CostTokenData;
-  activityHeatmap?: HeatmapData;
   sessions?: Session[];
 }
-
-const AGENT_COLORS = CHART_SERIES;
-const PROJECT_COLORS = CHART_SERIES;
-const getAgentColor = (i: number) => AGENT_COLORS[i % AGENT_COLORS.length];
-
-const PERIOD_OPTIONS = [
-  { value: '7d', label: 'Last 7 days' },
-  { value: '30d', label: 'Last 30 days' },
-  { value: '90d', label: 'Last 90 days' },
-  { value: '1y', label: 'Last year' },
-];
 
 const AGENT_LABELS: Record<string, { label: string; color: string }> = {
   claude_code: { label: 'Claude Code', color: 'orange' },
@@ -145,302 +59,6 @@ const SESSION_TYPE_LABELS: Record<string, string> = {
   auth_setup: 'Auth setup',
   tool_setup: 'Tool setup',
 };
-
-const chartTooltipStyle = {
-  backgroundColor: 'var(--app-bg-default)',
-  border: '1px solid var(--app-border-default)',
-  borderRadius: 8,
-  fontSize: 12,
-  color: 'var(--app-text-primary)',
-};
-
-function tickIntervalForPeriod(period: Period): number {
-  const days = period === '7d' ? 7 : period === '30d' ? 30 : period === '90d' ? 90 : 365;
-  return days <= 7 ? 0 : days <= 30 ? 4 : days <= 90 ? 9 : 29;
-}
-
-// --- Skeletons ---
-
-function SummarySkeletons() {
-  return (
-    <>
-      {Array.from({ length: 5 }).map((_, i) => (
-        <Paper key={i} withBorder p="lg" radius="md">
-          <Skeleton height={14} width={100} mb={12} />
-          <Skeleton height={38} width={90} />
-        </Paper>
-      ))}
-    </>
-  );
-}
-
-function ChartSkeleton({ height = 240 }: { height?: number }) {
-  return <Skeleton height={height} radius="sm" />;
-}
-
-// --- Panels ---
-
-function HeatmapPanel() {
-  const { activityHeatmap } = usePage<{ props: Props }>().props as unknown as Props;
-  if (!activityHeatmap) return null;
-
-  return (
-    <Paper withBorder p={24} radius="md" bg="var(--app-bg-card)" mb="xl">
-      <ContributionHeatmap days={activityHeatmap.days} />
-    </Paper>
-  );
-}
-
-function SummaryPanel() {
-  const { summary } = usePage<{ props: Props }>().props as unknown as Props;
-  if (!summary) return null;
-
-  const statBlocks = [
-    { label: 'Total Sessions', value: summary.totalSessions.toLocaleString(), icon: IconPlayerPlay, color: 'blue' },
-    { label: 'Total Cost', value: formatCostCents(summary.totalCostCents), icon: IconCoin, color: 'green' },
-    { label: 'Total Tokens', value: formatTokens(summary.totalTokens), icon: IconChartBar, color: 'violet' },
-    {
-      label: 'Avg Cost / Session',
-      value: formatCostCents(summary.avgCostCentsPerSession),
-      icon: IconClock,
-      color: 'orange',
-    },
-    { label: 'Workflows Run', value: summary.workflowsRun.toLocaleString(), icon: IconRoute, color: 'indigo' },
-  ];
-
-  return (
-    <>
-      {statBlocks.map((s) => (
-        <Paper key={s.label} withBorder p="lg" radius="md">
-          <Group gap={6} mb={12}>
-            <s.icon size={14} color={`var(--mantine-color-${s.color}-5)`} />
-            <Text size="xs" c="dimmed" tt="uppercase" style={{ letterSpacing: 0.5 }}>
-              {s.label}
-            </Text>
-          </Group>
-          <Text fw={700} lh={1.1} style={{ fontSize: 30 }}>
-            {s.value}
-          </Text>
-        </Paper>
-      ))}
-    </>
-  );
-}
-
-function ProjectBreakdownPanel() {
-  const { summary } = usePage<{ props: Props }>().props as unknown as Props;
-  if (!summary || !summary.projectBreakdowns || summary.projectBreakdowns.length === 0) return null;
-
-  const { projectBreakdowns } = summary;
-  const maxCost = projectBreakdowns.length > 0 ? Math.max(...projectBreakdowns.map((p) => p.costCents)) : 1;
-
-  return (
-    <Paper withBorder p={24} radius="md" mb="xl">
-      <Title order={4} mb={4}>
-        Per-Project Breakdown
-      </Title>
-      <Divider mb="md" />
-      <Box
-        style={{
-          display: 'grid',
-          gridTemplateColumns: '1fr 100px 130px 120px',
-          gap: 8,
-          padding: '8px 0',
-          borderBottom: '1px solid var(--app-border-default)',
-        }}
-      >
-        {['Project', 'Sessions', 'Cost', 'Tokens'].map((h) => (
-          <Text key={h} size="xs" c="dimmed" tt="uppercase" style={{ letterSpacing: 0.4 }}>
-            {h}
-          </Text>
-        ))}
-      </Box>
-      {projectBreakdowns.map((p, i) => {
-        const color = PROJECT_COLORS[i % PROJECT_COLORS.length];
-        const pct = maxCost > 0 ? (p.costCents / maxCost) * 100 : 0;
-        return (
-          <Box
-            key={p.projectId ?? 'none'}
-            style={{
-              display: 'grid',
-              gridTemplateColumns: '1fr 100px 130px 120px',
-              gap: 8,
-              alignItems: 'center',
-              padding: '12px 0',
-              borderBottom: i < projectBreakdowns.length - 1 ? '1px solid var(--app-border-default)' : undefined,
-            }}
-          >
-            <Box>
-              <Group gap={8}>
-                <Box w={10} h={10} style={{ borderRadius: '50%', backgroundColor: color, flexShrink: 0 }} />
-                <Text size="sm" fw={500}>
-                  {p.projectName}
-                </Text>
-              </Group>
-              <Box
-                mt={4}
-                style={{ width: '80%', height: 4, borderRadius: 2, backgroundColor: 'var(--app-bg-elevated)' }}
-              >
-                <Box style={{ height: 4, borderRadius: 2, backgroundColor: color, width: `${pct}%` }} />
-              </Box>
-            </Box>
-            <Text size="sm" c="dimmed">
-              {p.sessions.toLocaleString()}
-            </Text>
-            <Text size="sm" fw={600}>
-              {formatCostCents(p.costCents)}
-            </Text>
-            <Text size="sm" c="dimmed">
-              {formatTokens(p.tokens)}
-            </Text>
-          </Box>
-        );
-      })}
-    </Paper>
-  );
-}
-
-function AgentActivityPanel() {
-  const { agentActivity } = usePage<{ props: Props }>().props as unknown as Props;
-  if (!agentActivity) return null;
-
-  const { sessionsByAgent } = agentActivity;
-
-  return (
-    <Paper withBorder p={24} radius="md" mb="xl">
-      <Title order={4} mb={4}>
-        Usage Breakdown by Agent Type
-      </Title>
-      <Divider mb="md" />
-      <Group gap="lg" align="center">
-        <Box style={{ width: '40%' }}>
-          <ResponsiveContainer width="100%" height={200}>
-            <PieChart>
-              <Pie
-                data={sessionsByAgent}
-                dataKey="sessions"
-                nameKey="agentType"
-                cx="50%"
-                cy="50%"
-                innerRadius={55}
-                outerRadius={85}
-                paddingAngle={3}
-              >
-                {sessionsByAgent.map((entry, idx) => (
-                  <Cell key={entry.agentType} fill={getAgentColor(idx)} />
-                ))}
-              </Pie>
-              <RechartsTooltip contentStyle={chartTooltipStyle} formatter={(v) => [`${Number(v)} sessions`]} />
-            </PieChart>
-          </ResponsiveContainer>
-        </Box>
-        <Box style={{ flex: 1 }}>
-          {sessionsByAgent.map((agent, idx) => (
-            <Group
-              key={agent.agentType}
-              gap="sm"
-              py={8}
-              style={{
-                borderBottom: idx < sessionsByAgent.length - 1 ? '1px solid var(--app-border-default)' : undefined,
-              }}
-            >
-              <Box w={10} h={10} style={{ borderRadius: '50%', backgroundColor: getAgentColor(idx), flexShrink: 0 }} />
-              <Box style={{ flex: 1 }}>
-                <Text size="xs" fw={500}>
-                  {agent.agentType}
-                </Text>
-                <Text size="xs" c="dimmed">
-                  {agent.sessions} sessions
-                </Text>
-              </Box>
-              <Text size="xs" fw={500} c="dimmed">
-                {formatCostCents(agent.costCents)}
-              </Text>
-            </Group>
-          ))}
-        </Box>
-      </Group>
-    </Paper>
-  );
-}
-
-function CostTokenPanel({ tickInterval }: { tickInterval: number }) {
-  const { costToken } = usePage<{ props: Props }>().props as unknown as Props;
-  if (!costToken) return null;
-
-  return (
-    <Grid mb="xl" gap="md">
-      <Grid.Col span={{ base: 12, md: 6 }}>
-        <Paper withBorder p={24} radius="md">
-          <Title order={4} mb={4}>
-            Daily Cost
-          </Title>
-          <Divider mb="md" />
-          <ResponsiveContainer width="100%" height={220}>
-            <AreaChart data={costToken.timeSeries} margin={{ top: 4, right: 8, bottom: 0, left: -10 }}>
-              <defs>
-                <linearGradient id="usage-grad-cost" x1="0" y1="0" x2="0" y2="1">
-                  <stop offset="5%" stopColor="var(--app-chart-1)" stopOpacity={0.3} />
-                  <stop offset="95%" stopColor="var(--app-chart-1)" stopOpacity={0} />
-                </linearGradient>
-              </defs>
-              <CartesianGrid strokeDasharray="3 3" stroke="var(--app-border-subtle)" />
-              <XAxis dataKey="date" tick={{ fontSize: 11 }} interval={tickInterval} />
-              <YAxis tick={{ fontSize: 11 }} tickFormatter={(v: number) => `$${(v / 100).toFixed(0)}`} />
-              <RechartsTooltip
-                contentStyle={chartTooltipStyle}
-                formatter={(v) => [formatCostCents(Number(v)), 'Cost']}
-              />
-              <Area
-                type="monotone"
-                dataKey="costCents"
-                name="Cost"
-                stroke="var(--app-chart-1)"
-                fill="url(#usage-grad-cost)"
-                strokeWidth={2}
-                dot={false}
-              />
-            </AreaChart>
-          </ResponsiveContainer>
-        </Paper>
-      </Grid.Col>
-      <Grid.Col span={{ base: 12, md: 6 }}>
-        <Paper withBorder p={24} radius="md">
-          <Title order={4} mb={4}>
-            Daily Token Consumption
-          </Title>
-          <Divider mb="md" />
-          <ResponsiveContainer width="100%" height={220}>
-            <AreaChart data={costToken.timeSeries} margin={{ top: 4, right: 8, bottom: 0, left: 0 }}>
-              <defs>
-                <linearGradient id="usage-grad-tokens" x1="0" y1="0" x2="0" y2="1">
-                  <stop offset="5%" stopColor="var(--app-chart-6)" stopOpacity={0.3} />
-                  <stop offset="95%" stopColor="var(--app-chart-6)" stopOpacity={0} />
-                </linearGradient>
-              </defs>
-              <CartesianGrid strokeDasharray="3 3" stroke="var(--app-border-subtle)" />
-              <XAxis dataKey="date" tick={{ fontSize: 11 }} interval={tickInterval} />
-              <YAxis tick={{ fontSize: 11 }} tickFormatter={(v: number) => formatTokens(v)} />
-              <RechartsTooltip
-                contentStyle={chartTooltipStyle}
-                formatter={(v) => [formatTokens(Number(v)), 'Tokens']}
-              />
-              <Area
-                type="monotone"
-                dataKey="totalTokens"
-                name="Tokens"
-                stroke="var(--app-chart-6)"
-                fill="url(#usage-grad-tokens)"
-                strokeWidth={2}
-                dot={false}
-              />
-            </AreaChart>
-          </ResponsiveContainer>
-        </Paper>
-      </Grid.Col>
-    </Grid>
-  );
-}
 
 function sessionTokenFmt(n: number): string {
   if (!n || n === 0) return '—';
@@ -544,7 +162,6 @@ const UsagePage = () => {
   // actually belongs to more than one, so single-company users see no change.
   const { currentUser } = usePage<SharedProps>().props;
   const companyName = (currentUser?.memberships?.length ?? 0) > 1 ? currentUser?.currentCompany?.name : null;
-  const tickInterval = useMemo(() => tickIntervalForPeriod(period), [period]);
 
   const navigate = (nextPeriod: string) => {
     router.get(
@@ -587,51 +204,7 @@ const UsagePage = () => {
           </Alert>
         )}
 
-        {/* Contribution heatmap */}
-        <Deferred data="activityHeatmap" fallback={<Skeleton height={140} radius="sm" mb="xl" />}>
-          <HeatmapPanel />
-        </Deferred>
-
-        {/* Summary stats */}
-        <SimpleGrid cols={{ base: 2, sm: 3, md: 5 }} mb="xl" spacing="md">
-          <Deferred data="summary" fallback={<SummarySkeletons />}>
-            <SummaryPanel />
-          </Deferred>
-        </SimpleGrid>
-
-        {/* Per-project breakdown */}
-        <Deferred data="summary" fallback={<Skeleton height={200} radius="sm" mb="xl" />}>
-          <ProjectBreakdownPanel />
-        </Deferred>
-
-        {/* Agent activity */}
-        <Deferred
-          data="agentActivity"
-          fallback={
-            <Box mb="xl">
-              <ChartSkeleton height={200} />
-            </Box>
-          }
-        >
-          <AgentActivityPanel />
-        </Deferred>
-
-        {/* Cost & token usage */}
-        <Deferred
-          data="costToken"
-          fallback={
-            <Grid mb="xl" gap="md">
-              <Grid.Col span={{ base: 12, md: 6 }}>
-                <ChartSkeleton height={220} />
-              </Grid.Col>
-              <Grid.Col span={{ base: 12, md: 6 }}>
-                <ChartSkeleton height={220} />
-              </Grid.Col>
-            </Grid>
-          }
-        >
-          <CostTokenPanel tickInterval={tickInterval} />
-        </Deferred>
+        <UsageAnalytics period={period} />
 
         {/* Sessions list */}
         <Deferred data="sessions" fallback={<Skeleton height={200} radius="sm" />}>
